Re-refactoring core/search into ia/IAEditionSearch
