+++ openlibrary/plugins/openlibrary/carousels.py
@@ -3,16 +3,20 @@
 from infogami.infobase.client import storify
 from infogami.utils.view import public
 
-from openlibrary.core import search
+from openlibrary.core.ia import IAEditionSearch
 from openlibrary.core import cache
 
 @public
 def get_editions_by_ia_query(query='', sorts=None, page=1, limit=None,
                              timeout=cache.DEFAULT_CACHE_LIFETIME):
+    def editions_by_ia_query(query='', sorts=None, page=1, limit=None):
+        # Enable method to be cacheable
+        if 'env' not in web.ctx:
+            delegate.fakeload()
+        editions = IAEditionSearch.get(query=query, sorts=sorts, page=page, limit=limit)
+        return editions
+
     results = cache.memcache_memoize(
-        search.editions_by_ia_query, 'editions.search_ia', timeout=timeout)(
+        editions_by_ia_query, 'editions.search_ia', timeout=timeout)(
             query=query, sorts=sorts, page=page, limit=limit)
     return storify(results)
-
-def setup():
-    pass
