+++ openlibrary/tests/core/test_ia.py
@@ -1,5 +1,6 @@
 from __future__ import print_function
 from openlibrary.core import ia
+from infogami import config
 
 def test_xml2dict():
     assert ia.xml2dict("<metadata><x>1</x><y>2</y></metadata>") == {"x": "1", "y": "2"}
@@ -40,3 +41,22 @@ def test_get_metaxml(monkeypatch, mock_memcache):
     # test with metadata errors
     metadata_json = "{}"
     assert ia.get_meta_xml("foo02bar") == {}
+
+def test_normalize_sorts():
+    sorts = ['loans__status__last_loan_date+desc']
+    assert(ia.IAEditionSearch._normalize_sorts(sorts) == ['loans__status__last_loan_date desc'])
+
+def test_ia_search_queries():
+    """Make sure IA Advanced Search and Human Browsable urls are constructed correctly"""
+    query = 'test'
+    prepared_query = 'mediatype:texts AND !noindex:* AND openlibrary_work:(*) AND loans__status__status:AVAILABLE AND test'
+    assert(ia.IAEditionSearch.MAX_LIMIT == 20)
+    advancedsearch_url = 'http://archive.org/advancedsearch.php?rows=20&q=mediatype%3Atexts+AND+%21noindex%3A%2A+AND+openlibrary_work%3A%28%2A%29+AND+loans__status__status%3AAVAILABLE+AND+test&output=json&fl%5B%5D=identifier&fl%5B%5D=loans__status__status&fl%5B%5D=openlibrary_edition&fl%5B%5D=openlibrary_work&page=1&sort%5B%5D='
+    browse_url = 'https://archive.org/search.php?query=mediatype:texts AND !noindex:* AND openlibrary_work:(*) AND loans__status__status:AVAILABLE AND test&sort='
+
+    _expanded_query = ia.IAEditionSearch._expand_api_query(query)
+    _params = ia.IAEditionSearch._clean_params(q=_expanded_query)
+    assert(_expanded_query == prepared_query)
+    assert(ia.IAEditionSearch._clean_params(q='test', limit=200).get('rows') == ia.IAEditionSearch.MAX_LIMIT)
+    assert(ia.IAEditionSearch._compose_advancedsearch_url(**_params) == advancedsearch_url)
+    assert(ia.IAEditionSearch._compose_browsable_url(prepared_query) == browse_url)
